Factor FileCell out of ValueCell
