+++ src/edu/colorado/csdms/wmt/client/ui/cell/ValueCell.java
@@ -1,22 +1,9 @@
 package edu.colorado.csdms.wmt.client.ui.cell;
 
-import com.google.gwt.core.shared.GWT;
-import com.google.gwt.dom.client.Style.Unit;
-import com.google.gwt.event.dom.client.ChangeEvent;
-import com.google.gwt.event.dom.client.ChangeHandler;
-import com.google.gwt.event.dom.client.ClickEvent;
-import com.google.gwt.event.dom.client.ClickHandler;
-import com.google.gwt.user.client.Window;
-import com.google.gwt.user.client.ui.Button;
-import com.google.gwt.user.client.ui.FormPanel;
-import com.google.gwt.user.client.ui.FormPanel.SubmitCompleteEvent;
 import com.google.gwt.user.client.ui.HorizontalPanel;
-import com.google.gwt.user.client.ui.ListBox;
 
-import edu.colorado.csdms.wmt.client.control.DataURL;
 import edu.colorado.csdms.wmt.client.data.ParameterJSO;
 import edu.colorado.csdms.wmt.client.ui.ParameterTable;
-import edu.colorado.csdms.wmt.client.ui.dialog.UploadDialogBox;
 
 /**
  * Used to display the value of a parameter in a {@link ParameterTable}, a
@@ -29,8 +16,6 @@ import edu.colorado.csdms.wmt.client.ui.dialog.UploadDialogBox;
 public class ValueCell extends HorizontalPanel {
 
   private ParameterJSO parameter;
-  private UploadDialogBox upload;
-  private ListBox fileDroplist;
 
   /**
    * Makes a ValueCell from the information contained in the input
@@ -48,12 +33,12 @@ public class ValueCell extends HorizontalPanel {
       return;
     }
 
-    // Make a cell to match the type -- choice, file or other.
+    // Add a cell to match the parameter type.
     String type = this.parameter.getValue().getType();
     if (type.matches("choice")) {
       this.add(new ChoiceCell(this));
     } else if (type.matches("file")) {
-      makeFileCell(this.parameter.getValue().getDefault());
+      this.add(new FileCell(this));
     } else if (type.matches("int")) {
       this.add(new IntegerCell(this));
     } else if (type.matches("float")) {
@@ -70,48 +55,6 @@ public class ValueCell extends HorizontalPanel {
   public void setParameter(ParameterJSO parameter) {
     this.parameter = parameter;
   }
-  
-  /**
-   * A worker that makes the {@link ValueCell} display a droplist and a file
-   * upload button for the "file" parameter type.
-   * 
-   * @param value the value of the parameter, a String
-   */
-  private void makeFileCell(String value) {
-    fileDroplist = new ListBox(false); // no multi select
-    fileDroplist.addChangeHandler(new ListSelectionHandler());
-    fileDroplist.setStyleName("wmt-DroplistBox");
-    fileDroplist.addStyleDependentName("upload");
-
-    // Load the droplist. If the value of the incoming parameter isn't listed
-    // in the component, append it to the end of the list and select it.
-    Integer nFiles = this.parameter.getValue().getFiles().length();
-    Integer selectedIndex = -1;
-    for (int i = 0; i < nFiles; i++) {
-      fileDroplist.addItem(this.parameter.getValue().getFiles().get(i));
-      if (fileDroplist.getItemText(i).matches(value)) {
-        selectedIndex = i;
-      }
-    }
-    if (selectedIndex > 0) {
-      fileDroplist.setSelectedIndex(selectedIndex);
-    } else {
-      fileDroplist.setSelectedIndex(fileDroplist.getItemCount() - 1);
-    }
-    fileDroplist.setVisibleItemCount(1); // show one item -- a droplist
-    this.add(fileDroplist);
-
-    Button uploadButton = new Button("<i class='fa fa-cloud-upload'></i>");
-    uploadButton.setStyleName("wmt-UploadButton");
-    uploadButton.addClickHandler(new UploadHandler());
-
-    uploadButton.setTitle("Upload file to server");
-    this.add(uploadButton);
-
-    this.setCellVerticalAlignment(fileDroplist, ALIGN_MIDDLE);
-    this.setCellVerticalAlignment(uploadButton, ALIGN_MIDDLE);
-    uploadButton.getElement().getStyle().setMarginLeft(3, Unit.PX);
-  }
 
   /**
    * Passes the modified value up to
@@ -124,94 +67,5 @@ public class ValueCell extends HorizontalPanel {
   public void setParameterValue(String value) {
     ParameterTable pt = (ParameterTable) ValueCell.this.getParent();
     pt.setValue(parameter, value);
-  }
-
-  /**
-   * A class to handle selection in the "choices" ListBox.
-   */
-  public class ListSelectionHandler implements ChangeHandler {
-    @Override
-    public void onChange(ChangeEvent event) {
-      GWT.log("(onChange)");
-      ListBox listBox = (ListBox) event.getSource();
-      String value = listBox.getValue(listBox.getSelectedIndex());
-      setParameterValue(value);
-    }
-  }
-
-  /**
-   * Handles a click on the Upload button.
-   */
-  public class UploadHandler implements ClickHandler {
-    @Override
-    public void onClick(ClickEvent event) {
-
-      ParameterTable pt = (ParameterTable) ValueCell.this.getParent();
-      if (!pt.data.modelIsSaved()) {
-        String msg =
-            "The model must be saved to the server"
-                + " before files can be uploaded.";
-        Window.alert(msg);
-        return;
-      }
-
-      upload = new UploadDialogBox();
-      upload.setText("Upload File...");
-
-      // Get the id of the model this file belongs to.
-      String modelId = ((Integer) pt.data.getMetadata().getId()).toString(); 
-      upload.getHidden().setValue(modelId);
-
-      // Where the form is to be submitted.
-      upload.getForm().setAction(DataURL.uploadFile(pt.data));
-      
-      upload.getForm().addSubmitCompleteHandler(new UploadCompleteHandler());
-      upload.center();
-    }
-  }
-
-  /**
-   * When the upload is complete and successful, add the name of the uploaded
-   * file to the {@link ValueCell} fileDroplist, select it, and save it as the
-   * value of this parameter.
-   */
-  public class UploadCompleteHandler implements FormPanel.SubmitCompleteHandler {
-    @Override
-    public void onSubmitComplete(SubmitCompleteEvent event) {
-      
-      upload.hide();
-      
-      if (event.getResults() != null) {        
-        
-        // Strip the fakepath from the filename.
-        String fileName =
-            upload.getUpload().getFilename().replace("C:\\fakepath\\", "");
-        
-        // Add the filename to the fileDroplist, but only if it's not there
-        // already.
-        Integer listIndex = -1;
-        for (int i = 0; i < fileDroplist.getItemCount(); i++) {
-          if (fileDroplist.getItemText(i).matches(fileName)) {
-            listIndex = i;
-          }
-        }
-        if (listIndex > 0) {
-          fileDroplist.setSelectedIndex(listIndex);
-        } else {
-          fileDroplist.addItem(fileName);
-          fileDroplist.setSelectedIndex(fileDroplist.getItemCount() - 1);
-        }
-        
-        // Like, important.
-        setParameterValue(fileName);
-        
-        // Say everything is alright.
-        Window.alert("File uploaded!");
-        
-        // Mark the model as unsaved.
-        ParameterTable pt = (ParameterTable) ValueCell.this.getParent();
-        pt.data.updateModelSaveState(false);
-      }
-    }
   }
 }
